+++ packages/moonstone/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact moonstone module, newest changes on the top.
 
+## unreleased
+
+### Added
+
+- `moonstone/Header` prop `hideLine` to hide `border-bottom`
+
 ## [3.0.0-beta.1] - 2019-07-15
 
 ### Removed